# api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_jcom-dev_zmanim_internal_models.ZmanimRequest:
    properties:
      date:
        description: YYYY-MM-DD format
        type: string
      elevation:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      publisher_id:
        type: string
      timezone:
        type: string
    type: object
  internal_handlers.APIError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  internal_handlers.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/internal_handlers.APIError'
      meta:
        $ref: '#/definitions/internal_handlers.ResponseMeta'
    type: object
...
  internal_handlers.FormulaDetails:
    properties:
      display_name:
        type: string
      explanation:
        type: string
      method:
        type: string
      parameters:
        additionalProperties: true
        type: object
    type: object
...
  internal_handlers.ResponseMeta:
    properties:
      request_id:
        type: string
      timestamp:
        type: string
    type: object
...
  internal_handlers.ZmanWithFormula:
    properties:
      formula:
        $ref: '#/definitions/internal_handlers.FormulaDetails'
      is_beta:
        type: boolean
      key:
        type: string
      name:
        type: string
      time:
        type: string
      time_category:
        type: string
    type: object
  internal_handlers.ZmanimLocationInfo:
    properties:
      city_id:
        type: string
      city_name:
        type: string
      country:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      region:
        type: string
      timezone:
        type: string
    type: object
  internal_handlers.ZmanimPublisherInfo:
    properties:
      id:
        type: string
      logo_url:
        type: string
      name:
        type: string
    type: object
  internal_handlers.ZmanimWithFormulaResponse:
    properties:
      cached:
        type: boolean
      cached_at:
        type: string
      date:
        type: string
      location:
        $ref: '#/definitions/internal_handlers.ZmanimLocationInfo'
      publisher:
        $ref: '#/definitions/internal_handlers.ZmanimPublisherInfo'
      zmanim:
        items:
          $ref: '#/definitions/internal_handlers.ZmanWithFormula'
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: support@zmanim.com
    name: Zmanim Lab Support
  description: API for Jewish prayer times calculation and publisher management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://zmanim.com/terms
  title: Zmanim Lab API
  version: "1.0"
paths:
...
  /publisher/cache:
    delete:
      description: Clears all cached zmanim calculations for the authenticated publisher
      parameters:
      - description: Publisher ID
        in: header
        name: X-Publisher-Id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Cache invalidated
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handlers.APIError'
              type: object
        "404":
          description: Publisher not found
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handlers.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: Invalidate publisher cache
      tags:
      - Publisher
...
  /zmanim:
    get:
      consumes:
      - application/json
      description: Calculates Jewish prayer times (zmanim) for a specific city and
        date, optionally using a publisher's custom algorithm
      parameters:
      - description: City ID from the cities database
        in: query
        name: cityId
        required: true
        type: string
      - description: Publisher ID for custom algorithm (uses default if not specified)
        in: query
        name: publisherId
        type: string
      - description: Date in YYYY-MM-DD format (defaults to today)
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Calculated zmanim with formula details
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handlers.ZmanimWithFormulaResponse'
              type: object
        "400":
          description: Invalid parameters
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handlers.APIError'
              type: object
        "404":
          description: City not found
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handlers.APIError'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handlers.APIError'
              type: object
      summary: Get zmanim for a city
      tags:
      - Zmanim
    post:
      consumes:
      - application/json
      description: Calculates zmanim using raw latitude/longitude coordinates. Prefer
        the GET /zmanim endpoint with cityId for better accuracy.
      parameters:
      - description: Coordinates and date
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jcom-dev_zmanim_internal_models.ZmanimRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Calculated zmanim
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handlers.ZmanimWithFormulaResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handlers.APIError'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handlers.APIError'
              type: object
      summary: Calculate zmanim by coordinates (legacy)
      tags:
      - Zmanim
securityDefinitions:
  BearerAuth:
    description: JWT Bearer token from Clerk authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"